Fixes arithmetic overflow

+++ src/LightningDB/EnvironmentInfo.cs
@@ -8,16 +8,16 @@ namespace LightningDB
         /// <summary>
         /// ID of the last used page 
         /// </summary>
-        public int LastPageNumber { get; set; }
+        public long LastPageNumber { get; set; }
 
         /// <summary>
         /// ID of the last committed transaction  
         /// </summary>
-        public int LastTransactionId { get; set; }
+        public long LastTransactionId { get; set; }
 
         /// <summary>
         /// Size of the data memory map 
         /// </summary>
-        public int MapSize { get; set; }
+        public long MapSize { get; set; }
     }
 }
